move FaraoImportException to crac-util
