+++ data/crac-creation/crac-creation-util/src/main/java/com/farao_community/farao/data/crac_creation/util/FaraoImportException.java
@@ -5,7 +5,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.farao_community.farao.data.crac_creation.creator.api;
+package com.farao_community.farao.data.crac_creation.util;
+
+import com.farao_community.farao.data.crac_creation.creator.api.ImportStatus;
 
 /**
  * @author Peter Mitri {@literal <peter.mitri at rte-france.com>}
